Introduce module initializers
