+++ src/dotnet/App.Server/AppHost.cs
@@ -1,4 +1,4 @@
-using ActualChat.Hosting;
+using ActualChat.App.Server.Initializers;
 using ActualChat.Mesh;
 using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.Configuration.Json;
@@ -79,60 +79,48 @@ public class AppHost : IDisposable
             || type?.IsAssignableTo(typeof(IAsyncDisposable)) == true ? type : null;
     }
 
-    public virtual async Task InvokeDbInitializers(CancellationToken cancellationToken = default)
+    public virtual async Task InvokeInitializers(CancellationToken cancellationToken = default)
+        => await InvokeInitializers([
+                new ExecuteDbInitializers(Services),
+                new ExecuteModuleInitializers(Services),
+            ],
+            cancellationToken
+        )
+        .ConfigureAwait(false);
+
+    private async Task InvokeInitializers(IEnumerable<IAggregateInitializer> initializers, CancellationToken cancellationToken = default)
     {
-        var hostInfo = Services.HostInfo();
 #if DEBUG
         // See MeshLockBase.DefaultLockOptions - locks expire much longer in DEBUG
         var mustLock = false;
 #else
+        var hostInfo = Services.HostInfo();
         var mustLock = hostInfo.IsTested;
 #endif
-        if (!mustLock) {
-            await InvokeDbInitializersUnsafe(cancellationToken).ConfigureAwait(false);
-            return;
-        }
 
+        var tasks = initializers.Select(initializer => mustLock
+            ? InvokeInitializersProtected(initializer, cancellationToken)
+            : InvokeInitializersUnsafe(initializer, cancellationToken));
+
+        await Task.WhenAll(tasks).ConfigureAwait(false);
+    }
+
+    private async Task InvokeInitializersProtected(IAggregateInitializer initializer, CancellationToken cancellationToken = default)
+    {
         var meshLocks = Services.MeshLocks<InfrastructureDbContext>().WithKeyPrefix(nameof(AppHost));
-        var lockHolder = await meshLocks.Lock(nameof(InvokeDbInitializers), "", cancellationToken).ConfigureAwait(false);
+        var lockKey = initializer.GetType().Name;
+        const string lockValue = nameof(InvokeInitializersProtected);
+        var lockHolder = await meshLocks.Lock(lockKey, lockValue, cancellationToken).ConfigureAwait(false);
         await using var _ = lockHolder.ConfigureAwait(false);
         var lockCts = cancellationToken.LinkWith(lockHolder.StopToken);
 
-        await InvokeDbInitializersUnsafe(lockCts.Token).ConfigureAwait(false);
+        await InvokeInitializersUnsafe(initializer, lockCts.Token).ConfigureAwait(false);
     }
 
-    public virtual async Task InvokeDbInitializersUnsafe(CancellationToken cancellationToken = default)
-    {
-        // InitializeSchema
-        await InvokeDbInitializers(
-            nameof(IDbInitializer.InitializeSchema),
-            (x, ct) => x.InitializeSchema(ct),
-            cancellationToken
-            ).ConfigureAwait(false);
-
-        // InitializeData
-        await InvokeDbInitializers(
-            nameof(IDbInitializer.InitializeData),
-            (x, ct) => x.InitializeData(ct),
-            cancellationToken
-            ).ConfigureAwait(false);
-
-        // RepairData
-        await InvokeDbInitializers(
-            nameof(IDbInitializer.RepairData),
-            x => x.ShouldRepairData,
-            (x, ct) => x.RepairData(ct),
-            cancellationToken
-            ).ConfigureAwait(false);
-
-        // VerifyData
-        await InvokeDbInitializers(
-            nameof(IDbInitializer.VerifyData),
-            x => x.ShouldVerifyData,
-            (x, ct) => x.VerifyData(ct),
-            cancellationToken
-            ).ConfigureAwait(false);
-    }
+    private static async Task InvokeInitializersUnsafe(IAggregateInitializer initializer, CancellationToken cancellationToken)
+        => await initializer
+            .InvokeAll(cancellationToken)
+            .ConfigureAwait(false);
 
     public virtual Task Run(CancellationToken cancellationToken = default)
         => Host.RunAsync(cancellationToken);
@@ -179,64 +167,4 @@ public class AppHost : IDisposable
         WebHostBuilderContext webHost,
         IServiceCollection services)
         => AppServicesBuilder?.Invoke(webHost, services);
-
-    private Task InvokeDbInitializers(
-        string name,
-        Func<IDbInitializer, CancellationToken, Task> invoker,
-        CancellationToken cancellationToken)
-        => InvokeDbInitializers(name, _ => true, invoker, cancellationToken);
-
-    private async Task InvokeDbInitializers(
-        string name,
-        Func<IDbInitializer, bool> mustInvokePredicate,
-        Func<IDbInitializer, CancellationToken, Task> invoker,
-        CancellationToken cancellationToken)
-    {
-        var log = Host.Services.LogFor(GetType());
-        var runningTaskSources = Host.Services.GetServices<IDbInitializer>()
-            .ToDictionary(x => x, _ => TaskCompletionSourceExt.New<bool>());
-        var runningTasks = runningTaskSources
-            .ToDictionary(kv => kv.Key, kv => (Task)kv.Value.Task);
-        foreach (var (dbInitializer, _) in runningTasks)
-            dbInitializer.RunningTasks = runningTasks;
-        var tasks = runningTaskSources
-            .Select(kv => mustInvokePredicate.Invoke(kv.Key) ? InvokeOne(kv.Key, kv.Value) : Task.CompletedTask)
-            .ToArray();
-
-        try {
-            await Task.WhenAll(tasks).ConfigureAwait(false);
-        }
-        finally {
-            foreach (var (dbInitializer, _) in runningTasks)
-                dbInitializer.RunningTasks = null!;
-        }
-        return;
-
-        async Task InvokeOne(IDbInitializer dbInitializer, TaskCompletionSource<bool> initializedSource)
-        {
-            using var _ = dbInitializer.Activate();
-            var dbInitializerName = $"{dbInitializer.GetType().GetName()}.{name}";
-            try {
-                using var _1 = Tracer.Default.Region(dbInitializerName);
-                log.LogInformation("{DbInitializer} started", dbInitializerName);
-                var task = invoker.Invoke(dbInitializer, cancellationToken);
-                if (task.IsCompletedSuccessfully)
-                    log.LogInformation("{DbInitializer} completed synchronously (skipped?)", dbInitializerName);
-                else {
-                    await task.ConfigureAwait(false);
-                    log.LogInformation("{DbInitializer} completed", dbInitializerName);
-                }
-                initializedSource.TrySetResult(default);
-            }
-            catch (OperationCanceledException) {
-                initializedSource.TrySetCanceled(cancellationToken);
-                throw;
-            }
-            catch (Exception e) {
-                log.LogError(e, "{DbInitializer} failed", dbInitializerName);
-                initializedSource.TrySetException(e);
-                throw;
-            }
-        }
-    }
 }
