+++ src/dotnet/App.Server/Initializers/ExecuteModuleInitializers.cs
@@ -1,0 +1,11 @@
+using ActualChat.Hosting;
+
+namespace ActualChat.App.Server.Initializers;
+
+public class ExecuteModuleInitializers(IServiceProvider services): IAggregateInitializer
+{
+    public async Task InvokeAll(CancellationToken cancellationToken)
+        => await Task.WhenAll(services.GetServices<IModuleInitializer>()
+                .Select(instance => instance.Initialize(cancellationToken))
+            ).ConfigureAwait(false);
+}
