+++ src/dotnet/App.Server/Initializers/IAggregateInitializer.cs
@@ -1,0 +1,6 @@
+namespace ActualChat.App.Server.Initializers;
+
+public interface IAggregateInitializer
+{
+    Task InvokeAll(CancellationToken cancellationToken);
+}
