+++ src/dotnet/Core/Hosting/IModuleInitializer.cs
@@ -1,0 +1,6 @@
+namespace ActualChat.Hosting;
+
+public interface IModuleInitializer
+{
+    Task Initialize(CancellationToken cancellationToken);
+}
